requirements.txt
# =============================================
# Core Web Framework
# =============================================
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.9
aiofiles==24.1.0

# =============================================
# Database & Migrations
# =============================================
sqlalchemy==2.0.36
alembic==1.13.3
aiosqlite==0.20.0
psycopg2-binary==2.9.9

# =============================================
# Document Processing
# =============================================
PyMuPDF==1.25.1  # Fast PDF extraction - imports as 'fitz'
python-docx==1.1.2
python-magic==0.4.27; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'
Pillow==10.4.0

# =============================================
# AI / LLM Clients
# =============================================
openai==1.51.0
anthropic==0.34.0
groq==0.11.0
together==1.2.1

# =============================================
# Text-to-Speech & Audio
# =============================================
gtts==2.5.3
pydub==0.25.1
pyttsx3==2.97
edge-tts==6.1.12

# =============================================
# Authentication & Security
# =============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jwt==4.1.0

# =============================================
# Configuration & Utilities
# =============================================
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
email-validator==2.2.0
loguru==0.7.2
watchfiles==0.24.0

# =============================================
# Testing
# =============================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# =============================================
# Development & Code Quality
# =============================================
black==24.8.0
isort==5.13.2
ruff==0.6.9
mypy==1.11.2
pre-commit==3.8.0

# =============================================
# System Dependencies (Install separately)
# =============================================
# Ubuntu/Debian: sudo apt install tesseract-ocr poppler-utils
# macOS: brew install tesseract poppler
# Windows: Install Tesseract from https://github.com/UB-Mannheim/tesseract/wiki